# openapi/controlplane.openapi.yaml
openapi: 3.1.0
info:
  title: Evo Control Plane API
  version: 0.2.0
servers:
  - url: http://127.0.0.1:8080
security:
  - bearerAuth: []
paths:
...
  /v1/runs:
    get:
      operationId: listRuns
      parameters:
        - $ref: "#/components/parameters/WorkspaceIDQuery"
      responses:
        "200":
          description: Run list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RunListResponse"
    post:
      operationId: createRun
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateRunRequest"
      responses:
        "201":
          description: Created run
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RunEnvelopeResponse"
  /v1/runs/{runID}:
    get:
      operationId: getRun
      parameters:
        - $ref: "#/components/parameters/RunIDPath"
      responses:
        "200":
          description: Run envelope
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RunEnvelopeResponse"
        "404":
          $ref: "#/components/responses/ErrorResponse"
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: EVO
  parameters:
    WorkspaceIDQuery:
      in: query
      name: workspace_id
      required: true
      schema:
        type: string
...
    RunIDPath:
      in: path
      name: runID
      required: true
      schema:
        type: string
...
  responses:
    ErrorResponse:
      description: Error response
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
  schemas:
...
    ErrorResponse:
      type: object
      required: [error]
      properties:
        error:
          type: string
...
    ConfigMap:
      type: object
      additionalProperties: true
...
    RunbookStep:
      type: object
      required: [slug, kind]
      properties:
        slug: { type: string }
        kind: { type: string }
    Runbook:
      type: object
      required:
        - slug
        - title
        - family
        - description
        - tool_connections
        - approval_required
        - expected_artifacts
        - failure_modes
        - rollback_notes
        - steps
      properties:
        slug: { type: string }
        title: { type: string }
        family: { type: string }
        description: { type: string }
        tool_connections:
          type: array
          items: { type: string }
        approval_required: { type: boolean }
        expected_artifacts:
          type: array
          items: { type: string }
        failure_modes:
          type: array
          items: { type: string }
        rollback_notes: { type: string }
        steps:
          type: array
          items:
            $ref: "#/components/schemas/RunbookStep"
...
    TaskRun:
      type: object
      required:
        - id
        - org_id
        - workspace_id
        - runbook_slug
        - status
        - current_step
        - approval_state
        - requested_by_surface
        - requested_by_agent
        - context
        - created_at
        - updated_at
      properties:
        id: { type: string }
        org_id: { type: string }
        workspace_id: { type: string }
        environment_id: { type: string }
        runbook_slug: { type: string }
        status: { type: string }
        current_step: { type: integer }
        approval_state: { type: string }
        requested_by_surface: { type: string }
        requested_by_agent: { type: string }
        context:
          $ref: "#/components/schemas/ConfigMap"
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
    Artifact:
      type: object
      required:
        - id
        - run_id
        - workspace_id
        - kind
        - content_type
        - storage_key
        - created_by_surface
        - created_by_agent
        - created_at
      properties:
        id: { type: string }
        run_id: { type: string }
        workspace_id: { type: string }
        kind: { type: string }
        content_type: { type: string }
        storage_key: { type: string }
        created_by_surface: { type: string }
        created_by_agent: { type: string }
        created_at:
          type: string
          format: date-time
    ApprovalRequest:
      type: object
      required:
        - id
        - run_id
        - workspace_id
        - status
        - reason
        - requested_by_surface
        - requested_by_agent
        - created_at
      properties:
        id: { type: string }
        run_id: { type: string }
        workspace_id: { type: string }
        status: { type: string }
        reason: { type: string }
        decision_note: { type: string }
        requested_by_surface: { type: string }
        requested_by_agent: { type: string }
        created_at:
          type: string
          format: date-time
        decided_at:
          type: string
          format: date-time
    AuditEvent:
      type: object
      required:
        - id
        - org_id
        - workspace_id
        - kind
        - message
        - payload
        - created_at
      properties:
        id: { type: string }
        org_id: { type: string }
        workspace_id: { type: string }
        run_id: { type: string }
        approval_request_id: { type: string }
        kind: { type: string }
        message: { type: string }
        payload:
          $ref: "#/components/schemas/ConfigMap"
        created_at:
          type: string
          format: date-time
...
    RunEnvelope:
      type: object
      required: [run, approvals, artifacts, events]
      properties:
        run:
          $ref: "#/components/schemas/TaskRun"
        approvals:
          type: array
          items:
            $ref: "#/components/schemas/ApprovalRequest"
        artifacts:
          type: array
          items:
            $ref: "#/components/schemas/Artifact"
        events:
          type: array
          items:
            $ref: "#/components/schemas/AuditEvent"
        runbook:
          $ref: "#/components/schemas/Runbook"
...
    CreateRunRequest:
      type: object
      required: [workspace_id, runbook_slug]
      properties:
        workspace_id: { type: string }
        environment_id: { type: string }
        runbook_slug: { type: string }
        context:
          $ref: "#/components/schemas/ConfigMap"
...
    RunEnvelopeResponse:
      type: object
      required: [item]
      properties:
        item:
          $ref: "#/components/schemas/RunEnvelope"
...
    RunListResponse:
      type: object
      required: [items]
      properties:
        items:
          type: array
          items:
            $ref: "#/components/schemas/TaskRun"
...